Add worker-src and wasm-eval to CSP meta tag
By setting to Content-Security-Policy HTTP header, Bay of Plenty can
make the CSP rules only stricter than what is specfied in the meta tag.
It would have to remove the meta tag in order to relax the rules.
By default, be specifying only a she in script-src, web assembly gets
disabled. This is fixed by adding 'wasm-eval'.
Also, when no worker-src is specified, worker-src defaults to
script-src, which in out case prevents workers from being created
(because they don't match the sha of the app script)
This is fixed by explicitly allowing workers to be loaded from blobs
and data URLs.